requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

sqlalchemy==2.0.32

pdfplumber==0.11.4
pypdf==4.3.1

numpy==1.26.4
faiss-cpu==1.8.0.post1

sentence-transformers==3.0.1

requests==2.32.3
streamlit==1.37.1

python-dotenv==1.0.1
tqdm==4.66.4
